+++ src/model/commands/SetPalette.java
@@ -3,7 +3,8 @@ package model.commands;
 import model.turtle.Turtle;
 
 public class SetPalette extends Command{
-	public double execute(Turtle t, int index, double r, double g, double b) {
+	public double execute(Turtle t, double index, double r, double g, double b) {
+		
 		return 0;
 	}
